third_party/WebKit/LayoutTests/inspector-protocol/input/dispatchKeyEvent-focus-expected.txt


focus foo
blur foo
focus bar
blur bar
focus baz
================
value of foo:foo
value of bar:bar
value of baz:
